fix(cli): align resume EventStore resolution
